Enable soft-fp on i586
This is a porting of the patches in meta-zephyr-sdk to gcc-8.2.0

We merge:
	recipes-devtools/gcc/files/libgcc-t-zephyr.patch
	recipes-devtools/gcc/files/config-x86-multi-gcc.patch
	recipes-devtools/gcc/files/x86-multilib.patch
	recipes-devtools/gcc/files/config-libgcc-softfp.patch

Into one patch.  There are some changes to libgcc/config.host in
gcc-8.2.0 that required updating the patch to that file.  There's some
common code to handle soft-fp on x86.  In gcc-8.2.0 that common code
would by default get applied to a zephyr build.  However the common
code doesn't handle soft-fp the way we need so we add
i[34567]86-zephyr-elf as a special case similar to
i[34567]86-*-elfiamcu.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>
Signed-off-by: Anas Nashif <anas.nashif@intel.com>