Clang 9 bug workaround
Clang 9 has a bug in which it reports a `GCC_VERSION` greater than or
equal to 300 yet does not implement `__builtin_constant_p` as does GCC.

This workaround ensures that `__builtin_constant_p` is not used when
building with Clang 9.

For more details, refer to the linked GCC bug [1].

Note that this workaround is required for building GCC 10.3 on the
macOS 11 and above.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=92061

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>