RISC-V: Add base instruction set multilibs
This commit adds the following base instruction set multilibs that can
be used for every practical extension permutation:

* rv32i_zicsr_zifencei
* rv32e_zicsr_zifencei
* rv64i_zicsr_zifencei

These base instruction set multilibs are mapped to the compatible
extension permutations that do not have a dedicated multilib in order
to increase the ISA coverage of the toolchain.

Note that the Zicsr and Zifencei extensions are still specified for the
base instruction set multilibs because the Zephyr RISC-V architecture
port requires them and it is not practical to configure a RISC-V core
that does not support these instruction sets.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>