RISC-V: Build RV64 multilib with mcmodel=medany
This commit updates the Zephyr RISC-V multilib configurations to
explicitly specify `mcmodel=medany` when building the RV64 multilibs
because the GCC defaults to the `medlow` code model when the `mcmodel`
flag is not specified.

Note that the Zephyr build system builds the RV64 targets using the
`medany` code model because Zephyr uses the "flat" memory model and
the addressing capabilities of the `medlow` model is too limited for
it.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>