RISC-V: Use RV64 mcmodel=medany multilibs when mcmodel is not specified
This commit updates the multilib configurations such that the multilibs
that are built for the `medany` code model are used when the `mcmodel`
flag is not explicitly specified (i.e. when `medlow` code model is in
use).

Note that the multilibs compiled with the `medany` code model can be
used with the application code compiled with the `medlow` code model.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>