microblaze: Fix -Os right shift optimization is allowed into delay slot
During picolibc testing, it's found that `-Os` produces assembly code that
compiler squeezes into a single delay slot. Thus, only the first
instruction (the one in delay slot) emitted by this optimization is
executed and the rest is skipped.

This is a regression introduced by applying Microblaze gcc patches #24.
This patch is a 14.3.0 equivalent of #37.

Signed-off-by: Alp Sayin <alpsayin@gmail.com>