arm: Add missing multilib default values
The arm multilib configuration includes two more parameters which
affect multilib selection, marm/mthumb and mfloat-abi. Without those,
the default multilib selection is mis-specified and the only reason it
works is because '.' is the fall-back path.

Add "marm" and "mfloat-abi=soft" to MULTILIB_DEFAULTS to actually
match when the compiler is run without any target parameters.

This hasn't caused any problems in practice because there are no
non-default multilib options which can be applied to the default
-march target as it has neither an FPU nor any branch protection
support. Specifying another cpu or architecture always sets -marm and
-mfloat-abi and so those multilib configuration don't rely on the
defaults.

Signed-off-by: Keith Packard <keithp@keithp.com>