Enable long path support on Windows host
Declare `longPathAware` in the GCC executable manifest to allow GCC
executables to address paths longer than `MAX_PATH`, which is 260
characters, on Windows hosts.

Note that this alone is not sufficient to enable long path support on
Windows; the corresponding registry key must also be enabled. For more
details, refer to the Microsoft documentation [1].

[1] https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>